requirements.txt
matplotlib~=3.1.3
numpy~=1.18.1
Pillow~=7.1.2
torch
torchvision~=0.4.2
tensorboard
future

tqdm~=4.46.0
scipy~=1.4.1